requirements.txt
fastapi
uvicorn
openai-whisper
torch
ffmpeg-python
python-multipart

